feat: allow multiple source/target sets (#3)
The config file can now be an array of config objects to maintain multiple
source->target file sets. For example, in a Keycloak installation, there are
separate properties files to hold the strings for different parts of the
login process, but we want to keep all their translations up to date.